replaced double with long (multiplied everything by 5)

+++ app/src/main/java/com/example/focustycoon/storage/UserRepository.kt
@@ -3,19 +3,19 @@ package com.example.focustycoon.storage
 import androidx.lifecycle.LiveData
 
 interface UserRepository {
-    fun getCurrentTokens(): LiveData<Double>
+    fun getCurrentTokens(): LiveData<Long>
     fun addTokens(duration: Long)
-    fun getConversionRate(): Double
-    fun getConversionRate(level: Int): Double
+    fun getConversionRate(): Long
+    fun getConversionRate(level: Int): Long
 
     fun getEfficiencyLevel(): LiveData<Int>
-    fun getEfficiencyUpgradeCost(): Double
-    fun getEfficiencyUpgradeCost(level: Int): Double
+    fun getEfficiencyUpgradeCost(): Long
+    fun getEfficiencyUpgradeCost(level: Int): Long
     fun upgradeEfficiency(): Boolean
 
     fun getCapacityLevel(): LiveData<Int>
     fun getMaxCapacity(): Int
-    fun getCapacityUpgradeCost(): Double
-    fun getCapacityUpgradeCost(level: Int): Double
+    fun getCapacityUpgradeCost(): Long
+    fun getCapacityUpgradeCost(level: Int): Long
     fun upgradeCapacity(): Boolean
 }
